feat: Add FpsCounter component ✨

+++ async/src/main.rs
@@ -11,24 +11,25 @@ pub mod tui;
 
 pub mod utils;
 
+use clap::Parser;
+use color_eyre::eyre::Result;
+
 use crate::{
   runner::Runner,
   utils::{initialize_logging, initialize_panic_handler, version},
 };
-use clap::Parser;
-use color_eyre::eyre::Result;
 
 //// ANCHOR: args
 // Define the command line arguments structure
 #[derive(Parser, Debug)]
 #[command(version = version(), about = "ratatui async template with crossterm and tokio")]
 struct Args {
-  /// Tick rate (ms)
-  #[arg(short, long, default_value_t = 250)]
-  tick_rate: usize,
-  /// Render tick rate (ms)
-  #[arg(short, long, default_value_t = 100)]
-  render_tick_rate: usize,
+  /// Tick rate (s)
+  #[arg(short, long, default_value_t = 1.0/4.0)]
+  tick_rate: f64,
+  /// Render tick rate (s)
+  #[arg(short, long, default_value_t = 1.0/60.0)]
+  render_tick_rate: f64,
 }
 //// ANCHOR_END: args
